+++ lib/widgets/surfaces/glass_bottom_bar.dart
@@ -551,7 +551,9 @@ class _GlassBottomBarState extends State<GlassBottomBar> {
                             glowBlurRadius: widget.glowBlurRadius,
                             glowSpreadRadius: widget.glowSpreadRadius,
                             glowOpacity: widget.glowOpacity,
-                            onTap: () => widget.onTabSelected(i),
+                            // onTap is null: all tap selection goes through
+                            // TabIndicator.onBarTapDown (prevents double-fire).
+                            onTap: null,
                           ),
                         ),
                       ),
@@ -619,7 +621,9 @@ class _GlassBottomBarState extends State<GlassBottomBar> {
                         glowBlurRadius: widget.glowBlurRadius,
                         glowSpreadRadius: widget.glowSpreadRadius,
                         glowOpacity: widget.glowOpacity,
-                        onTap: () => widget.onTabSelected(i),
+                        // onTap is null: all tap selection goes through
+                        // TabIndicator.onBarTapDown (prevents double-fire).
+                        onTap: null,
                       ),
                     ),
                   )